pyproject.toml
[tool.poetry]
name = "py-ldap-exim_mail_groups"
version = "0.1.0"
description = ""
authors = ["nsx.snx <nsx.snx@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.5"
python-ldap = "^3.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
